update the data type of numbers to float

+++ airbyte-integrations/connectors/source-google-ad-manager/source_google_ad_manager/data_classes.py
@@ -15,7 +15,7 @@ class CustomBaseModel(BaseModel):
             _type_: _description_
         """
         try:
-            value = int(value)
+            value = float(value)
         except (ValueError, TypeError):
             value = 0
         return value
